pyproject.toml
[project]
name = "nextword-horizon"
version = "0.1.0"
description = "Probabilistic Horizon Expansion for next-word prediction"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "gradio>=4.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "umap-learn>=0.5.4",
    "requests>=2.31.0",  # For RemoteHTTPAdapter
    "transformers>=4.35.0",  # For LocalHFAdapter
    "plotly>=5.17.0",  # For 3D/2D visualizations in UI
    "pandas>=2.0.0",  # For data tables in UI
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
local = [
    "torch>=2.0.0",  # For LocalHFAdapter (large dependency, optional)
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["horizon_core", "api", "ui"]

